pyproject.toml
[project]
name = "monitor"
version = "0.1.1"
description = ""
authors = [{ name = "Zhan Rongrui", email = "2742392377@qq.com" }]
dependencies = ["pyserial>=3.5", "loguru>=0.7.0"]
requires-python = "<3.13,>=3.11"
readme = "README.md"
license = { text = "MIT" }

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["F", "I", "B", "C4", "PGH", "W", "YTT"]
ignore = [
    "PGH003",
    "E501",
    "E741",
    "B008",
    "B904",
    "B023",
    "F841",
    "F821",
    "F841",
]

[tool.ruff.lint.isort]
known-first-party = ["maibobo"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "I002"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pdm.build]
excludes = ["./**/.git"]
includes = ["maibobo"]
source-includes = ["README.md"]
editable-backend = "path"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
